refactor: rewrite search & qrcode as a module
